requirements.txt
Flask
mss
Pillow
PyAutoGUI
qrcode
